Force all servers to respond to requests
Previously, secondaries could reject all requests and let the primary answer all
requests. This changes makes requests sticky: if a request cannot be completed,
the same request (and importantly, to the same server) will be retried. Over
time, this forces all secondaries to respond positively to requests in order to
continue to acrue points.