Add SPOF monkey
The SPOF monkey iterates through nodes in a random order, testing whether the
cluster can make progress if that node is split away from the rest of the
cluster.

This fixes an issue where we would accept a solution which was not resilient to
the failure of a single node.